# lib/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Repo Guardian API
  version: 0.1.0
  description: Canonical Milestone 7A contract for supervised repository analysis, security-hardened two-phase execution planning, durable execution history, and saved analysis runs. The public route surface remains `/api/analyze`, `/api/execution/plan`, `/api/execution/execute`, and `/api/runs*`, with additive read models under `/api/execution/plans/*`. All routes require mandatory Bearer token authorization using the `API_SECRET_KEY`.
servers:
  - url: /
paths:
  /api/analyze:
    post:
      operationId: analyzeRepository
      summary: Analyze a public GitHub repository snapshot
      x-repo-guardian-request-type: AnalyzeRepoRequest
      x-repo-guardian-response-type: AnalyzeRepoResponse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeRepoRequest"
      responses:
        "200":
          description: Repository analysis payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyzeRepoResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "429":
          $ref: "#/components/responses/ErrorResponse"
        "502":
          $ref: "#/components/responses/ErrorResponse"
  /api/execution/plan:
    post:
      operationId: createExecutionPlan
      summary: Request an approval-gated execution plan with a short-lived token
      x-repo-guardian-request-type: ExecutionPlanRequest
      x-repo-guardian-response-type: ExecutionPlanResponse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecutionPlanRequest"
      responses:
        "200":
          description: Planned execution actions with approval token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionPlanResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
  /api/execution/execute:
    post:
      operationId: executeExecutionPlan
      summary: Execute a planned set of actions by presenting a valid approval token
      x-repo-guardian-request-type: ExecutionExecuteRequest
      x-repo-guardian-response-type: ExecutionResult
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecutionExecuteRequest"
      responses:
        "200":
          description: Structured execution result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionResult"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
  /api/execution/plans/{planId}:
    get:
      operationId: getExecutionPlan
      summary: Read the durable execution plan detail and lifecycle state
      x-repo-guardian-response-type: ExecutionPlanDetailResponse
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Durable execution plan detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionPlanDetailResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
  /api/execution/plans/{planId}/events:
    get:
      operationId: listExecutionPlanEvents
      summary: Read ordered audit events for a durable execution plan
      x-repo-guardian-response-type: ExecutionPlanEventsResponse
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Durable execution plan audit events.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionPlanEventsResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
  /api/runs:
    get:
      operationId: listAnalysisRuns
      summary: List durably saved analysis runs
      x-repo-guardian-response-type: ListAnalysisRunsResponse
      responses:
        "200":
          description: Saved analysis run summaries.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListAnalysisRunsResponse"
    post:
      operationId: saveAnalysisRun
      summary: Save the current analysis payload as a durable run
      x-repo-guardian-request-type: SaveAnalysisRunRequest
      x-repo-guardian-response-type: SaveAnalysisRunResponse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveAnalysisRunRequest"
      responses:
        "201":
          description: Saved analysis run and summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveAnalysisRunResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
  /api/runs/{runId}:
    get:
      operationId: getAnalysisRun
      summary: Reopen a durably saved analysis run
      x-repo-guardian-response-type: GetAnalysisRunResponse
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Saved analysis run and summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAnalysisRunResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
  /api/runs/compare:
    post:
      operationId: compareAnalysisRuns
      summary: Compare two durably saved analysis runs
      x-repo-guardian-request-type: CompareAnalysisRunsRequest
      x-repo-guardian-response-type: CompareAnalysisRunsResponse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompareAnalysisRunsRequest"
      responses:
        "200":
          description: Saved analysis run comparison.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompareAnalysisRunsResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
components:
  responses:
    ErrorResponse:
      description: Structured error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        details:
          nullable: true
    AnalyzeRepoRequest:
      type: object
      required:
        - repoInput
      properties:
        repoInput:
          type: string
          minLength: 1
      x-repo-guardian-zod-schema: AnalyzeRepoRequestSchema
    AnalyzeRepoResponse:
      type: object
      description: Runtime schema is owned by AnalyzeRepoResponseSchema in @repo-guardian/shared-types.
      x-repo-guardian-zod-schema: AnalyzeRepoResponseSchema
    ExecutionPlanRequest:
      type: object
      description: Runtime schema is owned by ExecutionPlanRequestSchema in @repo-guardian/shared-types.
      x-repo-guardian-zod-schema: ExecutionPlanRequestSchema
    ExecutionPlanResponse:
      type: object
      description: Runtime schema is owned by ExecutionPlanResponseSchema in @repo-guardian/shared-types.
      x-repo-guardian-zod-schema: ExecutionPlanResponseSchema
    ExecutionExecuteRequest:
      type: object
      description: Runtime schema is owned by ExecutionExecuteRequestSchema in @repo-guardian/shared-types.
      x-repo-guardian-zod-schema: ExecutionExecuteRequestSchema
    ExecutionResult:
      type: object
      description: Runtime schema is owned by ExecutionResultSchema in @repo-guardian/shared-types.
      x-repo-guardian-zod-schema: ExecutionResultSchema
    ExecutionPlanDetailResponse:
      type: object
      description: Runtime schema is owned by ExecutionPlanDetailResponseSchema in @repo-guardian/shared-types.
      x-repo-guardian-zod-schema: ExecutionPlanDetailResponseSchema
    ExecutionPlanEventsResponse:
      type: object
      description: Runtime schema is owned by ExecutionPlanEventsResponseSchema in @repo-guardian/shared-types.
      x-repo-guardian-zod-schema: ExecutionPlanEventsResponseSchema
    SaveAnalysisRunRequest:
      type: object
      description: Runtime schema is owned by SaveAnalysisRunRequestSchema in @repo-guardian/shared-types.
      x-repo-guardian-zod-schema: SaveAnalysisRunRequestSchema
    SaveAnalysisRunResponse:
      type: object
      description: Runtime schema is owned by SaveAnalysisRunResponseSchema in @repo-guardian/shared-types.
      x-repo-guardian-zod-schema: SaveAnalysisRunResponseSchema
    ListAnalysisRunsResponse:
      type: object
      description: Runtime schema is owned by ListAnalysisRunsResponseSchema in @repo-guardian/shared-types.
      x-repo-guardian-zod-schema: ListAnalysisRunsResponseSchema
    GetAnalysisRunResponse:
      type: object
      description: Runtime schema is owned by GetAnalysisRunResponseSchema in @repo-guardian/shared-types.
      x-repo-guardian-zod-schema: GetAnalysisRunResponseSchema
    CompareAnalysisRunsRequest:
      type: object
      description: Runtime schema is owned by CompareAnalysisRunsRequestSchema in @repo-guardian/shared-types.
      x-repo-guardian-zod-schema: CompareAnalysisRunsRequestSchema
    CompareAnalysisRunsResponse:
      type: object
      description: Runtime schema is owned by CompareAnalysisRunsResponseSchema in @repo-guardian/shared-types.
      x-repo-guardian-zod-schema: CompareAnalysisRunsResponseSchema